requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Cors==4.0.0

python-dotenv==1.0.1
requests==2.31.0
razorpay==1.3.0
cryptography==42.0.5
Pillow==10.2.0
xhtml2pdf==0.2.8